+++ src/jk_git/__init__.py
@@ -2,7 +2,7 @@
 
 
 __author__ = "Jürgen Knauth"
-__version__ = "0.2022.4.10"
+__version__ = "0.2022.7.24"
 
 
 
@@ -15,3 +15,5 @@ from .GitWrapper import GitWrapper
 from .GitServerRepository import GitServerRepository
 from .GitWorkingCopy import GitWorkingCopy
 from .GitRemoteRepository import GitRemoteRepository
+
+